`continue` after many `if ... in line` lines (hopefully speed up)
Instead of going through each `if` statement for every line, this will
now go to the next line in the loop after a successful parsing.
I imagine this will work, but it could cause issues if multiple keywords
are on the same line. I don't expect this to be the case, but I don't
know that for sure.